Move tests to individual test folders
