+++ frontend/src/components/header/__tests__/Header.test.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import { ThemeProvider } from 'emotion-theming'
-import theme from '../theme'
-import { Header, H1, H2, H3, H4, H5, H6 } from '../components/header'
+import theme from '../../../theme'
+import { Header, H1, H2, H3, H4, H5, H6 } from '..'
 
 //these are just the same base tests as govuk package
