+++ frontend/src/components/ordered-list/__tests__/OrderedList.test.js
@@ -1,41 +1,40 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import { ThemeProvider } from 'emotion-theming'
-import theme from '../theme'
-import { UnorderedList, Ul } from '../components/unordered-list'
+import theme from '../../../theme'
+import { OrderedList, Ol } from '..'
 
-describe('Unordered List', () => {
+describe('Ordered List', () => {
   const example = 'example'
   const wrapper = (
     <ThemeProvider theme={theme}>
-      <UnorderedList
+      <OrderedList
         fontSize={[1, null, 2]}
         lineHeight={[1, null, 2]}
         pl={[5, null, 6]}
         mb={4}
       >
         <li>{example}</li>
-      </UnorderedList>
+      </OrderedList>
     </ThemeProvider>
   )
 
-  it('renders an UnorderedList and Ul component without crashing', () => {
+  it('renders an OrderedList and Ol without crashing', () => {
     const div = document.createElement('div')
     ReactDOM.render(wrapper, div)
     ReactDOM.render(
       <ThemeProvider theme={theme}>
-        <Ul>{example}</Ul>
+        <Ol>{example}</Ol>
       </ThemeProvider>,
       div,
     )
   })
 
-  it('UL accepts a listStyleType prop', () => {
+  it('OL accepts a listStyleType prop', () => {
     const div = document.createElement('div')
-    ReactDOM.render(wrapper, div)
     ReactDOM.render(
       <ThemeProvider theme={theme}>
-        <Ul listStyleType="square">{example}</Ul>
+        <Ol listStyleType="lower-roman">{example}</Ol>
       </ThemeProvider>,
       div,
     )
