+++ frontend/src/components/unordered-list/__tests__/UnorderedList.test.js
@@ -1,40 +1,41 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import { ThemeProvider } from 'emotion-theming'
-import theme from '../theme'
-import { OrderedList, Ol } from '../components/ordered-list'
+import theme from '../../../theme'
+import { UnorderedList, Ul } from '..'
 
-describe('Ordered List', () => {
+describe('Unordered List', () => {
   const example = 'example'
   const wrapper = (
     <ThemeProvider theme={theme}>
-      <OrderedList
+      <UnorderedList
         fontSize={[1, null, 2]}
         lineHeight={[1, null, 2]}
         pl={[5, null, 6]}
         mb={4}
       >
         <li>{example}</li>
-      </OrderedList>
+      </UnorderedList>
     </ThemeProvider>
   )
 
-  it('renders an OrderedList and Ol without crashing', () => {
+  it('renders an UnorderedList and Ul component without crashing', () => {
     const div = document.createElement('div')
     ReactDOM.render(wrapper, div)
     ReactDOM.render(
       <ThemeProvider theme={theme}>
-        <Ol>{example}</Ol>
+        <Ul>{example}</Ul>
       </ThemeProvider>,
       div,
     )
   })
 
-  it('OL accepts a listStyleType prop', () => {
+  it('UL accepts a listStyleType prop', () => {
     const div = document.createElement('div')
+    ReactDOM.render(wrapper, div)
     ReactDOM.render(
       <ThemeProvider theme={theme}>
-        <Ol listStyleType="lower-roman">{example}</Ol>
+        <Ul listStyleType="square">{example}</Ul>
       </ThemeProvider>,
       div,
     )
